+++ src/List_Edit.java
@@ -8,7 +8,7 @@ class List_Edit extends Math_Panel
 
     private JPanel exchanable, poly_panel;
     private Menu_Input[] poly_fields;
-    private int[] poly_array;
+    private double[] poly_array;
     private Menu_Label confirm_label;
 
     private boolean selected;
@@ -72,7 +72,7 @@ class List_Edit extends Math_Panel
         poly_panel.setBackground(new Color(204, 229, 255));
         poly_fields = new Menu_Input[number+1];
         Menu_Label[] poly_labels = new Menu_Label[number + 1];
-        poly_array = new int[number+1];
+        poly_array = new double[number+1];
 
         for (int i = number; i >= 0 ; i--)
         {
